+++ packages/core/src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './isLink';
 export * from './pathSchema';
 export * from './resolveHyperlinkPattern';
 export * from './patternUtils';
+export * from './sanitizeNodeProps';
